+++ packages/registry/src/index.ts
@@ -3,9 +3,10 @@ import {
   ActionType,
   type BridgeActionParams,
   type IActionPlugin,
-  type MintActionParams,
   PluginActionNotImplementedError,
+  type MintActionParams,
   type SwapActionParams,
+  type DelegateActionParams,
   type TransactionFilter,
 } from '@rabbitholegg/questdk'
 
@@ -18,6 +19,7 @@ import { Optimism } from '@rabbitholegg/questdk-plugin-optimism'
 import { Hop } from '@rabbitholegg/questdk-plugin-hop'
 import { Arbitrum } from '@rabbitholegg/questdk-plugin-arbitrum'
 import { GMX } from '@rabbitholegg/questdk-plugin-gmx'
+import { Tally } from '@rabbitholegg/questdk-plugin-tally'
 
 export const plugins: Record<string, IActionPlugin> = {
   [Connext.pluginId]: Connext,
@@ -29,6 +31,7 @@ export const plugins: Record<string, IActionPlugin> = {
   [Across.pluginId]: Across,
   [Optimism.pluginId]: Optimism,
   [GMX.pluginId]: GMX,
+  [Tally.pluginId]: Tally,
 }
 
 export const getPlugin = (pluginId: string) => {
@@ -43,7 +46,7 @@ export const executePlugin = (
   plugin: IActionPlugin,
   actionType: ActionType,
   params: ActionParams,
-): Promise<TransactionFilter> | Promise<PluginActionNotImplementedError> => {
+): Promise<TransactionFilter | PluginActionNotImplementedError> => {
   switch (actionType) {
     case ActionType.Bridge:
       return plugin.bridge(params as unknown as BridgeActionParams)
@@ -51,6 +54,11 @@ export const executePlugin = (
       return plugin.swap(params as unknown as SwapActionParams)
     case ActionType.Mint:
       return plugin.mint(params as unknown as MintActionParams)
+    case ActionType.Delegate: {
+      if (plugin.delegate === undefined) {
+        return Promise.reject(new PluginActionNotImplementedError())
+      } else return plugin.delegate(params as unknown as DelegateActionParams)
+    }
     default:
       throw new Error(`Unknown action type "${actionType}"`)
   }